% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{build_binary_mat}
\alias{build_binary_mat}
\title{Build a DE binary matrix from de_lists}
\usage{
build_binary_mat(
  de_lists,
  min_fc = NULL,
  max_pval = NULL,
  min_roc = NULL,
  max_bad_batches = 0
)
}
\arguments{
\item{de_lists}{The \code{de_lists} object.}

\item{min_fc}{The minimum fold change threshold}

\item{max_pval}{The maximum p-value threshold}

\item{min_roc}{The minimum ROC threshold}

\item{max_bad_batches}{The maximum number of batches that a gene
may not meet thresholds and still be called DE. Example: if 
max_bad_batches is 0, then the gene must meet these thresholds
in all batches to be called DE. If max_bad_batches is 1, then the
gene may fail the thresholds in 1 out of N batches and still be
called DE. Can also be a vector the same length as de_lists.}
}
\value{
The genes x metacluster binary matrix
}
\description{
Converts the \code{de_lists} object from Maggie's DE pipeline
to a genes x metaclusters binary matrix, where 1 indicates that
a gene is DE in that metacluster.
}
