% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_boxplot}
\alias{build_boxplot}
\title{Build a colData boxplot for a SingleCellExperiment}
\usage{
build_boxplot(
  sce_obj,
  x,
  y,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  y_transform = "identity",
  assay = "logcounts"
)
}
\description{
Build a boxplot for a continuous y variable in
the SingleCellExperiment object, stratified by a discrete
variable in the SingleCellExperiment colData.
}
