% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_categorical_color_palette}
\alias{build_categorical_color_palette}
\title{Build color palette for categorical cell types
 
Creates a collection of colors to plot categorical data,
where there shouldn't be any indication of order or scale.
Can theoretically provide an infinitely large color palette,
however there are losses. This method can provide up to
12 colors from a single palette. 13-17 colors come from
two palettes appended together. All 17 colors are technically
unique, but the design is not ideal and there are several
pairs that are difficult to distinguish. With 18 or more
colors, the 12-color palette simply repeats as necessary,
creating duplicate colors because in this intended use case,
(for a stacked violin plot), the violins will occur all in
a row and should be visually distinguishable, even if not
uniquely colored.}
\usage{
build_categorical_color_palette(n = NULL, colorgorical = FALSE)
}
\arguments{
\item{n}{Number of colors needed. If omitted, the max number
is returned.}

\item{colorgorical}{If TRUE, uses a longer palette from Colorgorical.
else, uses a modified combination of RColorBrewer Set1 and Set2.}
}
\value{
categorical color palette.
}
\description{
Build color palette for categorical cell types
 
Creates a collection of colors to plot categorical data,
where there shouldn't be any indication of order or scale.
Can theoretically provide an infinitely large color palette,
however there are losses. This method can provide up to
12 colors from a single palette. 13-17 colors come from
two palettes appended together. All 17 colors are technically
unique, but the design is not ideal and there are several
pairs that are difficult to distinguish. With 18 or more
colors, the 12-color palette simply repeats as necessary,
creating duplicate colors because in this intended use case,
(for a stacked violin plot), the violins will occur all in
a row and should be visually distinguishable, even if not
uniquely colored.
}
