% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_cluster_overlaps_df}
\alias{build_cluster_overlaps_df}
\title{Build dataframe to ggplot cluster overlap heatmap.}
\usage{
build_cluster_overlaps_df(
  sce1,
  sce2 = NULL,
  cluster = c("metacluster"),
  cell_id = c("Barcode"),
  verbose = TRUE
)
}
\arguments{
\item{sce1}{The first SingleCellExperiment containing
cell IDs and cluster labels for those cells in the
colData.}

\item{sce2}{The second SingleCellExperiment with the
same characteristics as sce1. If NULL, then all cells
in sce1 are used with two different clustering partitions
stored in the sce1 colData. If not NULL, then only the
cells with identical cell IDs are used.}

\item{cluster}{A 1 or 2-member vector containing the
colData column names for the cluster labels. If cluster
has length 1, the same column is used from both SCE
objects. If cluster has length 2, then \code{cluster[1]} 
is used for sce1 and likewise for 2.}

\item{cell_id}{A 1 or 2-member vector containing the
colData column names for the cell IDs. If cell_id
has length 1, the same column is used from both SCE
objects. If cell_id has length 2, then \code{cell_id[1]} 
is used for sce1 and likewise for 2.}

\item{verbose}{If TRUE, then the number of cells in each
SCE object and the number of cells in common is printed
to stdout.}
}
\value{
A data frame with 4 columns and c x d rows where
  c and d are the number of unique clusters in the two
  clustering partitions. The first 2 columns represent
  possible pairs of clusters, named "cluster_1" and "cluster_2".
  They refer to clusters in sce1 and sce2 or \code{cluster[1]}
  and \code{cluster[2]}. The 3rd and 4th columns contain
  the percentage of cluster_1 in cluster_2 or vice versa and
  are named "percent_1_in_2" and "percent_2_in_1".
}
\description{
Build a dataframe suitable for ggplot. The dataframe
allows plotting a heatmap showing the overlap of two
clustering partitions of the same cells. Each cell of
the heatmap is the percentage of the x-axis cluster
in the y-axis cluster. If 2 SCE objects are passed,
only the cells in common will be used.
}
