% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_hierarchical_color_palette}
\alias{build_hierarchical_color_palette}
\title{Build color palette for hierarchical cell types
 
For building a stacked violin plot where the cell types fall into
broad categories. For example, the first 3 cell types are all of
broad category 1, the next 2 are of broad category 2, and the last
is in broad category 3. In this case, the first three colors would
be shades of red, the second two would be shades of purple, and the
last would be a shade of orange. Thus the different broad categories
are visualizable, but the subtypes are also distinguishable. The
\code{mapping} argument in this case would be \code{c(1, 1, 1, 2, 2, 3)}.}
\usage{
build_hierarchical_color_palette(mapping)
}
\arguments{
\item{mapping}{numeric vector of broad category assignments for the
cell types.}
}
\value{
color palette for a sorted \code{mapping}.
}
\description{
Build color palette for hierarchical cell types
 
For building a stacked violin plot where the cell types fall into
broad categories. For example, the first 3 cell types are all of
broad category 1, the next 2 are of broad category 2, and the last
is in broad category 3. In this case, the first three colors would
be shades of red, the second two would be shades of purple, and the
last would be a shade of orange. Thus the different broad categories
are visualizable, but the subtypes are also distinguishable. The
\code{mapping} argument in this case would be \code{c(1, 1, 1, 2, 2, 3)}.
}
