% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_marker_dotplot}
\alias{build_marker_dotplot}
\title{Build an expression plot for markers in metaclusters.}
\usage{
build_marker_dotplot(
  sce_obj,
  markers,
  marker_names = NULL,
  cluster = "metacluster",
  plot_title = "Mean Expression and \% Expressed",
  stat = c("norm_mean", "z_score")
)
}
\description{
Build an expression plot for markers in metaclusters.
}
