% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_marker_expression_slides}
\alias{build_marker_expression_slides}
\title{Build two panel plots of marker expression and clusters
        
Creates and saves to disk a collection of plots, each a
2-panel plot of a 2-D embedding of a SingleCellExperiment
object. The left panel is colored by the expression of a
single gene, the right plot is colored by cluster for
easy comparison. All plots are saved as PNG files to
a given folder.}
\usage{
build_marker_expression_slides(
  sce_obj,
  markers,
  x,
  y,
  color = NULL,
  palette = NULL,
  group_labels = FALSE,
  group_min_size = 1,
  surface_density = FALSE,
  plot_title = NULL,
  plot_title_append = TRUE,
  legend_title = NULL,
  save = TRUE,
  path = NULL,
  w = 13.333,
  h = 7.5,
  return_plots = FALSE,
  order_files = FALSE,
  file_names = NULL,
  create_pdf = FALSE,
  keep_images = TRUE,
  pdf_title = "marker_expr.pdf"
)
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object to be plotted}

\item{markers}{A character vector of gene names. One 2-panel
plot will be generated per marker.}

\item{x}{Either a column name in \code{colData(sce_obj)} or
a numeric vector for the x-dimension of the 2-D embedding.}

\item{y}{Same as x, but the y-dimension.}

\item{color}{If null, all points will be black. If character string,
must be a column in \code{colData(sce_obj)} or a gene in
\code{rownames(sce_obj)}. If color is a gene, then expression will
be taken from the logcounts assay.}

\item{palette}{If NULL, will automatically pick colors with
\code{build_categorical_color_palette()}. Otherwise will use
palette (string vector of colors) to color points.}

\item{group_labels}{If TRUE, the right plot will have labels placed
at the mean position of each color group.}

\item{group_min_size}{The minimum size group in the right plot that
will be labeled or colored. Groups smaller than this will be
colored gray and unlabeled.}

\item{surface_density}{If \code{TRUE}, a surface density plot will
drawn instead of a marker expression plot on the left panel.}

\item{plot_title}{A string to be appended to the end of every
plot title. Can be a single static string, or a vector
of \code{length(markers)}. If \code{NULL}, the title will just be the
name of the marker gene.}

\item{plot_title_append}{If \code{TRUE}, then \code{plot_title} will be
appended to the end of the marker name. Otherwise, will replace the
marker name as the title. Can also be a single static string, or a vector
of \code{length(markers)}.}

\item{legend_title}{If \code{!NULL}, will be used as the legend title
for the right plot.}

\item{save}{If \code{TRUE}, will save all plots to disk
in the location indicated by \code{path}.}

\item{path}{A string indicating the folder in which the
PNG files will be saved. Must exist at time of function call.}

\item{w}{Width of plots in inches. Default is the size of a
widescreen PowerPoint slide.}

\item{h}{Height of plots in inches. Default is the size of a
widescreen PowerPoint slide.}

\item{return_plots}{If \code{TRUE}, will return the collection
of plots as a list. Otherwise returns \code{NULL}}

\item{order_files}{If \code{TRUE}, will add a 4 digit counter
to the front of each file name so that order is preserved.}

\item{create_pdf}{If \code{TRUE}, will also generate a PDF concatenating
all PNG files generated.}

\item{keep_images}{If \code{FALSE}, PNG files will be deleted after
concatenation into PDF file.}
}
\value{
A list of cowplot lists if return_plots is TRUE, NULL otherwise.
}
\description{
Build two panel plots of marker expression and clusters
        
Creates and saves to disk a collection of plots, each a
2-panel plot of a 2-D embedding of a SingleCellExperiment
object. The left panel is colored by the expression of a
single gene, the right plot is colored by cluster for
easy comparison. All plots are saved as PNG files to
a given folder.
}
