% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_stacked_violin_plot}
\alias{build_stacked_violin_plot}
\title{Build a stacked violin ggplot}
\usage{
build_stacked_violin_plot(
  sce_obj,
  marker_df,
  mapping = NULL,
  sort = TRUE,
  show_cluster = NULL,
  clusters = NULL,
  show_missing_clusters = FALSE
)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object holding the
expression values to be plotted in the \code{logcounts}
assay. The colData dataframe must have the cluster assignment
in the \code{metaclust} column.}

\item{marker_df}{Dataframe with 2 columns: \code{gene} and
\code{cluster}. All genes in this dataframe will be included
in the stacked violin plot.}

\item{mapping}{numeric vector of broad category assignments for the
cell types. If the cell types also have parent categories, this
vector should hold them. For example, for 5 cell types in 3
broad categories, \code{mapping} might be \code{c(1, 1, 2, 3, 3)}.}

\item{sort}{If true, then genes will be sorted according to cluster
with maximum mean expression.}

\item{show_cluster}{If NULL, then the gene names will be shown. If
!NULL, then the value of show_cluster will be printed after the
gene name, then the value of column 2 of marker_df will be appended
and used as row labels. Example: If show_cluster = ": Cluster ",
and column 2 of marker_df are the cluster names c(1:10), then
each row of the violins will be labeled "GENE: Cluster CLUSTERNAME".}

\item{clusters}{ordered vector of clusters to ensure consistent
coloring across multiple plots}

\item{show_missing_clusters}{If true, includes a single 0 point for
clusters that are not represented in the data. Method argument
clusters must not be NULL if show_missing_clusters is TRUE.}
}
\value{
A ggplot object holding the stacked violin plot.
}
\description{
Take a SingleCellExperiment object with expression values
in the \code{logcounts} assay, and a 2-column dataframe
of markers and generate a stacked violin plot, showing
the expression distributions per cluster per gene.
}
