% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build_thresholded_cluster_palette}
\alias{build_thresholded_cluster_palette}
\title{Build a categorical color palette for large clusters.}
\usage{
build_thresholded_cluster_palette(
  clusters,
  threshold = 5,
  omitted_color = "#000000"
)
}
\arguments{
\item{clusters}{A factor holding the cluster assignments
for each cell. The palette order corresponds to the factor
level order. This is ideally the \code{clusters} member of
a clustering results object.}

\item{threshold}{The minimum number of cells a cluster must
have to be assigned a color. Default is 5.}

\item{omitted_color.}{A hex string indicating the color to use
for clusters that don't meet the threshold. Default is black.}
}
\value{
A vector of hex colors the same length as the number
  of levels in \code{clusters}. The order of colors matches
  the order of the levels in \code{clusters}.
}
\description{
Builds a categorical color palette for clusters larger
than a certain threshold. Clusters smaller than the
threshold get colored black by default.
}
