% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_ari}
\alias{calc_ari}
\title{Calculate the Adjusted Rand Index for two clustering partitions.}
\usage{
calc_ari(clusters_1, clusters_2)
}
\arguments{
\item{clusters_1}{A vector of cluster labels indicating the first
partition to compare.}

\item{clusters_2}{A vector of cluster labels indicating the second
partition to compare.}
}
\value{
The Adjusted Rand Index of the two cluster partitions.
}
\description{
Calculate the Adjusted Rand Index for two clustering partitions,
returning a value that reflects the similarity of the clustering
partitions. The two vectors must be alternate clustering partitions
for the same entities.
}
