% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{calc_cluster_depth_de_vs_all_cells}
\alias{calc_cluster_depth_de_vs_all_cells}
\title{Calculate a "differential expression" test on read depth for clusters.}
\usage{
calc_cluster_depth_de_vs_all_cells(
  sce_obj,
  clust_ids,
  fc_thresh = 0.5,
  p_thresh = 1e-04,
  method = c("mw", "ttest"),
  missing_clusts = FALSE
)
}
\description{
Calculate a "differential expression" test on read depth for clusters.
}
