% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_qc.R
\name{calc_n_counts}
\alias{calc_n_counts}
\title{Calculate Library Size}
\usage{
calc_n_counts(sce_obj, assay = "counts")
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object.}

\item{assay}{The named assay to use.}
}
\value{
A numeric vector the same length as the number
  of cells in sce_obj. The number of counts for each cell.
}
\description{
Calculate a vector of number of counts for each cell.
}
