% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_qc.R
\name{calc_n_genes}
\alias{calc_n_genes}
\title{Calculate Number of Genes.}
\usage{
calc_n_genes(sce_obj, assay = "counts")
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object.}

\item{assay}{The named assay to use.}
}
\value{
A numeric vector the same length as the number
  of cells in sce_obj. The number of expressed genes for each cell.
}
\description{
Calculate a vector of number of non-zero genes for each cell.
}
