% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cluster_batch_bars}
\alias{cluster_batch_bars}
\title{Build a stratified bar plot showing cluster and batch abundance.}
\usage{
cluster_batch_bars(
  sce_obj,
  cluster = "metacluster",
  batch = "study_id",
  title = NULL
)
}
\description{
Build a stratified bar plot, showing the number of cells in
each cluster, with each bar's batch breakdown shown by
stratified fill.
}
