% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_infomap}
\alias{cluster_infomap}
\title{Perform Infomap Clustering.}
\usage{
cluster_infomap(
  pca_mat,
  k = 100,
  mutual = TRUE,
  n_trials = 10,
  process_communities = TRUE,
  name = "infomap"
)
}
\arguments{
\item{pca_mat}{The cell x PCA matrix.}

\item{k}{The number of neighbors to use in the nearest
neighbor graph.}

\item{mutual}{If TRUE, an undirected mutual nearest neighbors
graph will be constructed and used. If FALSE, a directed
nearest neighbors graph with no mutual constraint will be used.}

\item{n_trials}{Passed to the \code{nb.trials} argument of the
\code{igraph::cluster_infomap} function. Should be the number
of attempts to partition the network (any int >= 1).}

\item{process_communities}{If TRUE, the communities object
returned by the igraph clustering functions will be processed
into a custom list designed for Nathan's workflow.}

\item{name}{A string to be used for \code{name} if process_communities
is TRUE.}
}
\value{
If process_communities, a custom list containing 4 members.
  A vector of length n_cells containing the cluster assignments for
  the cells called "clusters". A vector of unique cluster names
  called "cluster_names". A string containing an optional name for
  this clustering result. A list of named properties, indicating the
  tunable parameters used for this particular clustering. If
  process_communities is \code{FALSE}, the igraph communities object
  is returned.
}
\description{
Perform clustering on a cell x PCA matrix using the
Infomap Community Detection algorithm.
}
