% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_leiden}
\alias{cluster_leiden}
\title{Perform Leiden Clustering.}
\usage{
cluster_leiden(
  pca_mat,
  k = 100,
  mutual = TRUE,
  process_communities = TRUE,
  name = "leiden",
  reindex = TRUE
)
}
\arguments{
\item{pca_mat}{The cell x PCA matrix.}

\item{k}{The number of neighbors to use in the nearest
neighbor graph.}

\item{mutual}{If TRUE, an undirected mutual nearest neighbors
graph will be constructed and used. If FALSE, a directed
nearest neighbors graph with no mutual constraint will be used.}

\item{process_communities}{If TRUE, the clusters returned by the
helper python script will be processed into a custom list designed
for Nathan's workflow.}

\item{name}{A string to be used for \code{name} if process_communities
is TRUE.}
}
\value{
If process_communities, a custom list containing 4 members.
  A vector of length n_cells containing the cluster assignments for
  the cells called "clusters". A vector of unique cluster names
  called "cluster_names". A string containing an optional name for
  this clustering result. A list of named properties, indicating the
  tunable parameters used for this particular clustering. If
  process_communities is \code{FALSE}, the vector of cluster
  assignments is returned.
}
\description{
Perform clustering on a cell x PCA matrix using the
Leiden Community Detection algorithm. Note that this
method is unlike the other clustering functions because
it has to interface with a helper python script. It is
much more fragile than the others, which are fully
implemented in R. This also means that it uses a vector
of cluster assignments, not an igraph communities object.
}
