% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_louvain}
\alias{cluster_louvain}
\title{Perform Louvain Clustering.}
\usage{
cluster_louvain(pca_mat, k = 100, process_communities = TRUE, name = "louvain")
}
\arguments{
\item{pca_mat}{The cell x PCA matrix.}

\item{k}{The number of neighbors to use in the nearest
neighbor graph.}

\item{process_communities}{If TRUE, the communities object
returned by the igraph clustering functions will be processed
into a custom list designed for Nathan's workflow.}

\item{name}{A string to be used for \code{name} if process_communities
is TRUE.}
}
\value{
If process_communities, a custom list containing 4 members.
  A vector of length n_cells containing the cluster assignments for
  the cells called "clusters". A vector of unique cluster names
  called "cluster_names". A string containing an optional name for
  this clustering result. A list of named properties, indicating the
  tunable parameters used for this particular clustering. If
  process_communities is \code{FALSE}, the igraph communities object
  is returned.
}
\description{
Perform clustering on a cell x PCA matrix using the
Louvain Community Detection algorithm.
}
