% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{combine_de_results}
\alias{combine_de_results}
\title{Combine batch-specific DE results.}
\usage{
combine_de_results(de_list, pval_col_pattern = "pval")
}
\arguments{
\item{pval_col_pattern}{Pattern to pass to grep for selecting pvalue
columns to combine. Empty string will combine all columns.}
}
\description{
Combine a de_list of clusters and batch-specific DE results into
one table per cluster, including a column of p-values combined
with Fisher's method.
}
