% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_qc.R
\name{filter_empty_drops}
\alias{filter_empty_drops}
\title{Filters empty droplets from droplet-based scRNA-seq data.}
\usage{
filter_empty_drops(
  sce_obj,
  min_read_count,
  assay = "counts",
  alpha = 0.05,
  keep = FALSE,
  marker_column = "non_empty",
  verbose = TRUE
)
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object.}

\item{min_read_count}{All cells with a library size lower than this
will be automatically filtered without calculation.}

\item{assay}{The SCE assay to use.}

\item{alpha}{The cutoff value for significance testing}

\item{keep}{If TRUE, the empty drops will be marked and kept instead
of being filtered out.}

\item{marker_column}{If keep == TRUE, the empty drops will be marked
in this column of \code{colData}.}

\item{verbose}{If TRUE, the statistics will be reported via stdout.}
}
\value{
The edited SCE object.
}
\description{
Uses the \code{DropletUtils::emptyDrops} function to filter
out empty droplets. Can just mark the empty drops in the 
\code{colData}, or can actually remove them.
}
