% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_dataframe}
\alias{generate_dataframe}
\title{Generate dataframe for feeding to ggplot}
\usage{
generate_dataframe(
  sce_obj,
  marker_df,
  sort = TRUE,
  show_cluster = NULL,
  clusters = NULL,
  show_missing_clusters = FALSE
)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object holding the
expression values to be plotted in the \code{logcounts}
assay. The colData dataframe must have the cluster assignment
in the \code{metaclust} column.}

\item{marker_df}{Dataframe with 2 columns: \code{gene} and
\code{cluster}. All genes in this dataframe will be included
in the stacked violin plot.}

\item{sort}{If true, then genes will be sorted according to cluster
with maximum mean expression.}

\item{show_cluster}{If NULL, then the gene names will be shown. If
!NULL, then the value of show_cluster will be printed after the
gene name, then the value of column 2 of marker_df will be appended
and used as row labels. Example: If show_cluster = ": Cluster ",
and column 2 of marker_df are the cluster names c(1:10), then
each row of the violins will be labeled "GENE: Cluster CLUSTERNAME".}

\item{clusters}{ordered vector of clusters to ensure consistent
coloring across multiple plots}

\item{show_missing_clusters}{If true, includes a single 0 point for
clusters that are not represnted in the data. Method argument
clusters must not be NULL if show_missing_clusters is TRUE.}
}
\value{
A new dataframe ready to be plugged into ggplot.
}
\description{
Take a SingleCellExperiment object and a dataframe
listing markers to display and construct a dataframe
that can be fed to ggplot for building a stacked violin
plot.
}
