% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{metadata_histogram}
\alias{metadata_histogram}
\title{Plot a histogram of a variable in rowData and colData.}
\usage{
metadata_histogram(
  sce_obj,
  x,
  row_metadata = FALSE,
  title = NULL,
  x_label = NULL,
  bins = NULL,
  binwidth = NULL,
  mean = TRUE,
  median = TRUE
)
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object.}

\item{x}{The name of the variable to plot.}

\item{row_metadata}{TRUE if x is in rowData, FALSE if in colData.}

\item{title}{The title of the plot.}

\item{x_label}{The x-axis label of the plot.}

\item{bins}{The number of bins to use. Cannot be set if binwidth is set.}

\item{binwidth}{The size of each bin. Cannot be set if bins is set.}

\item{mean}{If TRUE, a line will be plotted to indicate the mean of the
distribution.}

\item{median}{If TRUE, a line will be plotted to indicate the median of
the distribution.}
}
\description{
Plots a ggplot histogram of a metadata variable. Can also
annotate with the mean and/or median. Note that the mean and
median legend values will only be accurate to 0 decimal places.
}
