% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{pca_elbow_plot}
\alias{pca_elbow_plot}
\title{Plot a PCA Screeplot}
\usage{
pca_elbow_plot(pca_results, npcs = 50)
}
\arguments{
\item{pca_results}{Anything with an sdev vector in it. Typically
from \code{prcomp}, \code{princomp}, or \code{rsvd:rpca}.}

\item{npcs}{The number of PCs to plot.}
}
\value{
Nothing. It just plots the plot.
}
\description{
Plots the principal components against their variances
to create an "elbow plot" or a screeplot. This plot is typically
used to estimate the number of PCs to include.
}
