% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_expression_density_estimate}
\alias{plot_expression_density_estimate}
\title{Plot a kernel density estimate for summed gene expression over cells}
\usage{
plot_expression_density_estimate(
  sce_obj,
  genes = NULL,
  cells = NULL,
  plot_title = "",
  x_title = "Summed Gene Expression",
  y_title = "",
  assay = "logcounts"
)
}
\description{
Plot a kernel density estimate for summed gene expression over cells
}
