% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_qc.R
\name{plot_read_count_cell_ranks}
\alias{plot_read_count_cell_ranks}
\title{Plot rank-ordered library size.}
\usage{
plot_read_count_cell_ranks(sce_obj, assay = "counts")
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object.}

\item{assay}{The SCE assay to use.}
}
\value{
Nothing. Just plots the plot.
}
\description{
Plots the library size of each cell, rank-ordered by decreasing
library size. Also adds the knee and inflection points.
}
