% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print_clusters}
\alias{print_clusters}
\title{Print a summary of a clustering results object.}
\usage{
print_clusters(cluster_results, max_print = 50)
}
\arguments{
\item{cluster_results}{A clustering results object produced
by \code{process_igraph_communities}.}

\item{max_print}{If there are more clusters than this number,
the individual clusters will not be printed.}
}
\value{
Nothing. Just prints to stdout.
}
\description{
Prints a summary of a clustering results object produced
by \code{process_igraph_communities}.
}
