% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{process_igraph_communities}
\alias{process_igraph_communities}
\title{Process an igraph Communities Object.}
\usage{
process_igraph_communities(comm, name = NULL, properties = NULL)
}
\arguments{
\item{comm}{The igraph communities object.}

\item{name}{The name attribute to be used for the clustering results object.}

\item{properties}{The list of clustering parameters to be stored in the properties
attribute of the clustering results object.}
}
\value{
A list with 4 members. clusters = the vector of cluster assignments
  for all cells. cluster_names = the vector of unique cluster names. 
  name = a string naming this object. properties = a list of named parameters
  intended to store the parameters used in this particular clustering.
}
\description{
Process an igraph communities object into a custom list containing
clustering results for Nathan's pipeline.
}
