% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{read_de_results}
\alias{read_de_results}
\title{Read a folder of csv files generated by write.de()}
\usage{
read_de_results(
  folder,
  batch_prefix = "batch_",
  cluster_prefix = "cluster_",
  batch_index = 2,
  cluster_index = 4
)
}
\arguments{
\item{folder}{The name of the folder holdin the csv
files. It must not contain anything else. Additionally
the files must be named in the style that write.de()
writes them, i.e. "batch_B_cluster_C_de.csv" where
B and C are the batch and cluster number.}

\item{batch_prefix}{A character prefix to prepend to the
batch number when the names are being generated for
the elements of the sub lists.}

\item{cluster_prefix}{A character prefix to prepend to the
cluster number when the names are being generated for the
top list.}

\item{batch_index}{The index of the batch number in the file
name, when the file name is split by "_".}

\item{cluster_index}{The index of the cluster number in the
file name, when the file name is split by "_".}
}
\value{
A list of lists of dataframes. The top list has
  C elements where C is the number of metaclusters. Each
  element is a sub list. Each sub list has B elements where
  B is the number of batches with cells in that cluster. Each
  sub list's element is a data frame containing the DE
  analysis results for the cells in that cluster in that
  batch. The results are mean_exp, log2fc, de_pval, de_roc,
  pct_in, and pct_out for each gene.
}
\description{
Reads a folder of csv files generated by write.de()
into a list of lists of dataframes containing the
DE marker analysis results. The top list is sorted
by cluster. Then each sub list is sorted by batch.
Each sub list contains the DE analysis results
for that cluster in that batch.
}
