% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reindex_clusters}
\alias{reindex_clusters}
\title{Reindex cluster labels to 1:N.}
\usage{
reindex_clusters(clusters)
}
\arguments{
\item{clusters}{Vector of numeric cluster labels. Can be numeric,
character, or factor.}
}
\value{
A numeric vector of the same length as \code{clusters}
  containing the reindexed cluster labels.
}
\description{
Reindexes a vector of cluster labels to the labels 1:N where N
is the number of unique clusters. Use to clean up clusters
that have been subset and skip labels. Must be used on numeric
cluster labels.
}
