% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{run_fast_tsne}
\alias{run_fast_tsne}
\title{Runs FIt-SNE on an scRNA-seq PCA matrix.}
\usage{
run_fast_tsne(mat, n_pcs = NULL, fast_tsne_path = NULL)
}
\arguments{
\item{mat}{A cells x PCs matrix.}

\item{n_pcs}{The number of PCs to include.}

\item{fast_tsne_path}{The path to the fast_tsne executable.
Not technically necessary as long as the copy of fast_tsne.R
is appropriately altered for this package.}
}
\value{
A cells x 2 matrix holding the t-SNE coordinates.
}
\description{
Runs the FIt-SNE version of t-SNE on a cells x PCs matrix
from a scRNA-seq dataset. Implements the recommendations made
by Kobak & Berens (10.1038/s41467-019-13056-x) for t-SNE on
scRNA-seq data.
}
