% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{run_umap}
\alias{run_umap}
\title{Runs UMAP on an scRNA-seq PCA matrix.}
\usage{
run_umap(mat, n_pcs = NULL)
}
\arguments{
\item{mat}{A cells x PCs matrix.}

\item{n_pcs}{The number of PCs to include.}
}
\value{
A cells x 2 matrix holding the UMAP coordinates.
}
\description{
Runs the umap version of UMAP on a cells x PCs matrix
from an scRNA-seq dataset.
}
