% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatter_plot_sce}
\alias{scatter_plot_sce}
\title{Build a colored scatterplot ggplot object}
\usage{
scatter_plot_sce(
  sce_obj,
  x,
  y,
  color = NULL,
  na_trans = TRUE,
  na_val = "#000000",
  size = 0.5,
  cont_color_trans = "identity",
  palette = NULL,
  cont_palette = ggplot2::scale_colour_distiller(palette = "Blues", direction = 1,
    trans = cont_color_trans),
  alpha_highlight = NULL,
  alpha_min = 0.3,
  plot_title = "Cells",
  legend_title = NULL,
  show_legend = TRUE,
  group_labels = FALSE,
  group_min_size = 1,
  surface_density = FALSE,
  shuffle = FALSE
)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object to be plotted.}

\item{x}{x coordinates for plotting. Can be a character string in
\code{colnames(colData(sce_obj))} or a numeric vector with the
same length as \code{dim(sce_obj)[2]}.}

\item{y}{y coordinates for plotting. Can be a character string in
\code{colnames(colData(sce_obj))} or a numeric vector with the
same length as \code{dim(sce_obj)[2]}.}

\item{color}{If null, all points will be black. If character string,
must be a column in \code{colData(sce_obj)} or a gene in
\code{rownames(sce_obj)}. If color is a gene, then expression will
be taken from the logcounts assay.}

\item{na_trans}{If color is a discrete scale, this is passed to na.trans
in \code{ggplot2::scale_colour_manual()}.}

\item{na_val}{If color is a discrete scale, this is passed to na.value
in \code{ggplot2::scale_colour_manual()}.}

\item{size}{Will be passed to ggplot2::geom_point(ggplot2::aes_string)}

\item{cont_color_trans}{trans argument for the continuous color distiller
scale used by ggplot. Example: pass "log10" to do a log10 transformation
on the continuous color variable.}

\item{palette}{If NULL, will automatically pick colors with
\code{build_categorical_color_palette()}. Otherwise will use
palette (string vector of colors) to color points.}

\item{cont_palette}{A continuous color palette compatible with
ggplot2. Default is the Blues palette using the distiller function.}

\item{alpha_highlight}{A list of 2 character vectors. The first
character vector must be of length 1 and hold a column name from
\code{colData(sce_obj)}. The second character vector must be
a vector of factor levels in the column indicated by the first
vector. These levels will be plotted at full intensity. All
other levels will be plotted at \code{alpha_min} transparency.}

\item{alpha_min}{The transparency alpha value that will be used
for all levels of \code{alpha_highlight[[1]][1]} not indicated
by \code{alpha_highlight[[2]]}.}

\item{plot_title}{Character string for plot title. "Cells" by default.}

\item{legend_title}{Character string for legend title.
Default is \code{color}.}

\item{show_legend}{Logical indicating whether or not to include a legend.}

\item{group_labels}{If \code{TRUE}, the right plot will have labels
placed at the mean position of each color group.}

\item{group_min_size}{The minimum size group in the right plot that
will be labeled or colored. Groups smaller than this will be
colored gray and unlabeled.}

\item{surface_density}{If \code{TRUE}, a surface density plot will be
plotted instead of a marker expression color. Only compatible when
color is a marker gene.}

\item{shuffle}{If \code{TRUE}, the cells will be plotted in random order.}
}
\value{
A ggplot object holding the scatter plot.
}
\description{
Create a 2-D scatterplot from a SingleCellExperiment object.
Intended for plotting low-dimensional embeddings of scRNA-seq
data, colored by batch or cluster or gene expression.
}
