% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scran_hvg}
\alias{scran_hvg}
\title{Returns the scran-chosen highly variable genes from the SCE object.}
\usage{
scran_hvg(sce_obj, assay = "logcounts", alpha = 0.5)
}
\arguments{
\item{sce_obj}{The SingleCellExperiment object.}

\item{assay}{The name of the assay to be used.}

\item{alpha}{The FDR cutoff. Any gene with an FDR <= to this
value will be labeled highly variable.}
}
\description{
Uses \code{scran::modelGeneVar} to select highly variable
genes from the SCE object. Then returns them as a vector.
}
