% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sort_genes}
\alias{sort_genes}
\title{Generate sorted expression matrix for stacked violin plot}
\usage{
sort_genes(sce_obj, marker_df)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object holding the
expression values to be plotted in the \code{logcounts}
assay. The colData dataframe must have the cluster assignment
in the \code{metaclust} column.}

\item{marker_df}{Dataframe with 2 columns: \code{gene} and
\code{cluster}. All genes in this dataframe will be included
in the stacked violin plot.}
}
\value{
An expression matrix containing only the marker genes, sorted
  so that the first genes are markers for the first cluster.
}
\description{
Given a SingleCellExperiment and a list of markers, it sorts
the markers by "cluster with maximum expression" and returns
an expression matrix with only those markers in the sorted order.
This means that the top gene will be maximally expressed in the first
cluster and so on.
}
